+++ tests/test_utils.py
@@ -129,6 +129,9 @@ class TestUtils(absltest.TestCase):
                 <geom name="b4/g1" type="sphere" size=".1" mass=".1"/>
               </body>
             </body>
+            <body name="geomless">
+              <inertial pos="0 0 0" mass=".1" diaginertia="1 1 1"/>
+            </body>
           </worldbody>
         </mujoco>
         """
@@ -137,7 +140,65 @@ class TestUtils(absltest.TestCase):
         actual_geom_ids = utils.get_subtree_geom_ids(model, b1_id)
         geom_names = ["b1/g1", "b1/g2", "b2/g1"]
         expected_geom_ids = [model.geom(g).id for g in geom_names]
-        self.assertListEqual(actual_geom_ids, expected_geom_ids)
+        self.assertSetEqual(set(actual_geom_ids), set(expected_geom_ids))
+        b3_id = model.body("b3").id
+        actual_geom_ids = utils.get_subtree_geom_ids(model, b3_id)
+        geom_names = ["b3/g1", "b4/g1"]
+        expected_geom_ids = [model.geom(g).id for g in geom_names]
+        self.assertSetEqual(set(actual_geom_ids), set(expected_geom_ids))
+        geomless_id = model.body("geomless").id
+        actual_geom_ids = utils.get_subtree_geom_ids(model, geomless_id)
+        self.assertListEqual(actual_geom_ids, [])
+        world_id = 0
+        actual_geom_ids = utils.get_subtree_geom_ids(model, world_id)
+        expected_geom_ids = [i for i in range(model.ngeom)]
+        self.assertSetEqual(set(actual_geom_ids), set(expected_geom_ids))
+
+    def test_get_subtree_body_ids(self):
+        xml_str = """
+        <mujoco>
+          <worldbody>
+            <body name="b1" pos=".1 -.1 0">
+              <joint type="free"/>
+              <geom name="b1/g1" type="sphere" size=".1" mass=".1"/>
+              <geom name="b1/g2" type="sphere" size=".1" mass=".1" pos="0 0 .5"/>
+              <body name="b3">
+                <joint type="hinge" range="0 1.57" limited="true"/>
+                <geom name="b3/g1" type="sphere" size=".1" mass=".1"/>
+                <body name="b4" pos="1 1 1">
+                    <geom name="b4/g1" type="sphere" size=".1" mass=".1"/>
+                </body>
+              </body>
+              <body name="b2" pos="1 1 1">
+                <geom name="b2/g1" type="sphere" size=".1" mass=".1"/>
+              </body>
+            </body>
+            <body name="b5" pos="1 1 1">
+              <joint type="free"/>
+              <geom name="b5/g1" type="sphere" size=".1" mass=".1"/>
+              <body name="b6">
+                <joint type="hinge" range="0 1.57" limited="true"/>
+                <geom name="b6/g1" type="sphere" size=".1" mass=".1"/>
+              </body>
+            </body>
+          </worldbody>
+        </mujoco>
+        """
+        model = mujoco.MjModel.from_xml_string(xml_str)
+        b1_id = model.body("b1").id
+        actual_body_ids = utils.get_subtree_body_ids(model, b1_id)
+        body_names = ["b1", "b3", "b4", "b2"]
+        expected_body_ids = [model.body(b).id for b in body_names]
+        self.assertSetEqual(set(actual_body_ids), set(expected_body_ids))
+        b5_id = model.body("b5").id
+        actual_body_ids = utils.get_subtree_body_ids(model, b5_id)
+        body_names = ["b5", "b6"]
+        expected_body_ids = [model.body(b).id for b in body_names]
+        self.assertSetEqual(set(actual_body_ids), set(expected_body_ids))
+        world_id = 0
+        actual_body_ids = utils.get_subtree_body_ids(model, world_id)
+        expected_body_ids = [i for i in range(model.nbody)]
+        self.assertSetEqual(set(actual_body_ids), set(expected_body_ids))
 
 
 if __name__ == "__main__":