+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Added `utils.get_subtree_body_ids` to get all bodies belonging to the subtree starting at a given body.
+  - Example usage of this function can be found in the [ALOHA example script](examples/arm_aloha.py) where it is used to selectively apply gravity compensation torques to the left and right arm bodies.
+
 ### Changed
 
 - Improved ALOHA example script.